pyproject.toml
[project]
name = "llm-explorations"
version = "0.1.0"
description = "Explorations with LLMs"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.0.1",
    "anthropic>=0.34",
    "ell-ai[all]>=0.0.14",
    "ipykernel>=6.29.5",
    "matplotlib>=3.9.2",
    "numpy>=2.1.2",
    "openai>=1.51.2",
    "pandas>=2.2.3",
    "pathlib>=1.0.1",
    "pillow>=10.4.0",
    "pytest>=8.3.3",
    "python-dotenv",
    "replicate>=1.0.2",
    "requests>=2.32.3",
    "tabulate>=0.9.0",
    "torch>=2.4.1",
    "torchaudio>=2.4.1",
    "torchvision>=0.19.1",
    "transformers>=4.45.2",
]
